Initial commit of the new FxgConverter8
The FxgConverter8 is a tool to convert Adobe FXG files to either JavaFX 8 controls (control, skin, behavior and css or extend region and css) or HTML5 Canvas controls. If a file was dropped to the "drop zone" a FileSystemWatcher will update the preview as soon as the dropped fxg file was modified. This version is still work in progress but does work...